Implement SmptAuthEmailSender and GmailSmtpEmailSender
There have been two modifications to the EmailSender interface:
(1) I introduced a new functional interface MessageEditor. It sounds like the
    MessageAuthor that we removed before and it is indeed quite similar with 1
    modification: it declares Exception so it can pass on all checked exceptions
    that get thrown during message creation - particularly MessagingException
    which is thrown by all(?) methods of MimeMessage. Please feel free to
    propose an alternative solution or an alternative name.
(2) Also due to the issue with MimeMessage-methods throwing MessagingException,
    EmailSender.send(...) now throws EmailSenderException (e.g. those that
    wrap MessagingException during message creation). This seems necessary
    because the message shall be created before EmailSender.send(...) returns,
    i.e. these exceptions cannot(?) be wrapped inside the CompletionStage that
    is used for sending the email.

I decided to make all internal methods of SmtpAuthEmailSender protected because
I myself never like to come across classes whose authors could not think of any
reason to subclass them. One particular reason to subclass SmtpAuthEmailSender
and to override SmtpAuthEmailSender.createProperties(...) is that there are many
configuration properties in the Java Mail API that make it likely that the
current implementation does not fit all SMTP servers in the world.

In the JavaDoc of SmtpAuthEmailSender.createProperties(...) I wrote "Consult the
source code of this method before you override it" instead of listing the
properties the method sets. I'm not sure whether you deploy the source code in
Maven along with the binaries. I thought not replicating the properties in the
JavaDoc might make it less likely that the JavaDoc becomes outdated.