Adjust exception documentation for EmailSender.send(...)
The example in EmailSender.send(...) did not correctly represent that the EmailDeliveryException is
now contained in the CompletionStage directly without being wrapped in a CompletionException. This
change occurred when we switched from using CompletableFuture.supplyAsync(...) to Executor.exec(...)
and the direct use of CompletionStage.execeptionally(...).

I also took up a change in the @return JavaDoc that Laura proposed.